First Project in CISCO packet tracer
Decided to actually make a project to design a scenario for a LAN and create it in CISCO packet tracer. Trying to learn the various things that I've been studying through out the Network+ book. This is only the physical layout of the network more or less, no real configurations so far.